+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+- Use a thread local instead of rwlock to store the spans
+- Drop opentelemetry::http as it was pulling the `blocking` feature of reqwest and causing hanging issues
+- Fixes a runtime panic when determining the span duration
+
 ## [0.11.0]
 
 -   Drop unused dependencies (including the worker crate).